pull into SupervisorCard
